Move Tree::edit implementation to another file
